separate nyx driver from std driver
